Fixed the last 6 warnings, enabled "Treat warnings as errors"
- Fixed incompatible code property of ISpdyError with the base class NSError: replaced the getter declaration by a property one, using NS_ENUM to make ISpdyErrorCode compatible with NSInteger (the type of NSError's code property), and updated a format string due to this change (common.m:185)
- Fixed _closeSocket declaration and implementation differing in return type
- Fixed an integer precision warning (ispdy.m:1047)
- Fixed the null argument passed to non-null parameter warning in loop.m, added for this a dummy timer method and passing that to the timer
- Fixed retain cycle warning by using weak reference to self in request.m
- Updated the project file and the podspec to enable the "Treat warnings as errors" build setting

+++ src/common.m
@@ -181,8 +181,8 @@
     case kISpdyErrSendAfterClose:
       return @"ISpdy error: request sent after close";
     default:
-      return [NSString stringWithFormat: @"Unexpected spdy error %d",
-          self.code];
+      return [NSString stringWithFormat: @"Unexpected spdy error %ld",
+          (long)self.code];
   }
 }
